Peer_review fixes
Added some fixes based on the feedback from the peer review comments I
recieved.
- Added a BoatDAL.cs file that handles the boat.txt file
- Added a MemberDAL.cs file that handles the member.txt file
- Changed the way a memberList is shown to the user, it now uses just
one method instead of two, regardless of if it's verbose or compact.
This is true for both user.cs and console.cs
- Changed a confusing code snippet relating to adding a new member.
Previously, the code made it seem that the member was in fact not saved
the way one could think it would be. It's now using memberID instead of
a member object.